feat(03-03): extract api-core.ts into 6 focused modules (R12)
- Created api-headers.ts: single source of truth for Cache-Control and rate-limit headers
- Created api-request.ts: canonical getClientIdentifier, getRequestMetadata, parseRequestBody
- Created api-rate-limit.ts: checkRateLimitOrRespond, RateLimitPresets, RateLimitConfig
- Created api-csrf.ts: validateCSRFOrRespond integrating csrf-protection
- Created api-pagination.ts: parsePaginationParams, createPaginationMeta, PaginationParams
- Created api-logging.ts: logApiRequest, logApiResponse
- Cleaned api-utils.ts: removed all duplicated functions (createResponseHeaders, getClientIdentifier, getRequestMetadata, parseRequestBody, logApiRequest, logApiResponse)
- Deleted api-core.ts (618 LOC god-file eliminated)
